refactor: Remove IGptDecoderBatched interface
- Removed the IGptDecoderBatched interface and its references.
- Moved decoder_batch::Input class to gptDecoderBatched.h.
- Updated includes in relevant files to reflect the new structure.
- Adjusted method signatures in GptDecoderBatched to remove interface dependencies.
